pyproject.toml
[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai-sync"
version = "1.2.0"
description = "Self-contained AI config sync for Codex, Cursor, Gemini, and Claude Code."
readme = "README.md"
authors = ["Loup Peluso <loup.peluso@gmail.com>"]
license = "LicenseRef-PolyForm-Noncommercial-1.0.0"
homepage = "https://github.com/Ephasme/ai-sync"
repository = "https://github.com/Ephasme/ai-sync"
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
packages = [{ include = "ai_sync", from = "src" }]
include = [
    "LICENSE",
    "src/ai_sync/.client-versions.json",
    "src/ai_sync/web/static/**/*",
]

[tool.poetry.urls]
Issues = "https://github.com/Ephasme/ai-sync/issues"

[tool.poetry.dependencies]
python = ">=3.11"
onepassword-sdk = ">=0.4.0"
pyyaml = ">=6.0"
pydantic = ">=2,<3"
jsonpointer = ">=3.0.0"
tomli = ">=2.0"
tomli-w = ">=1.0"
rich = ">=13.0"
pathspec = ">=0.12"
dependency-injector = "^4.48.3"
fastapi = "^0.135.1"
uvicorn = {extras = ["standard"], version = "^0.41.0"}

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
pyright = ">=1.1.390"
ruff = ">=0.9.0"
pre-commit = ">=4.0"
httpx = "^0.28.1"

[tool.poetry.scripts]
ai-sync = "ai_sync.cli:main"

[tool.coverage.run]
branch = true
source = ["ai_sync"]

[tool.coverage.report]
show_missing = true
skip_covered = false
omit = ["tests/*"]
